[#61] Temporary fix for BrowserSync on Node 0.12.x.
This commit vendors the npm module from:
mwbrooks/docpad-plugin-browsersync#3-fix-node-0-12-x

We vendor it instead of using a Github dependency reference because the
module requires the CoffeeScript to be compiled and it is not handled by
a npm post-install hook. So for now, we will vendor the compiled assets
until the pull request is resolved.