[#78] Unpublish content except Getting Started and Hello World.
In issue #78, we discussed adding a 'draft' or 'published' meta-data
option. However, DocPad already supports a 'write' meta-data option
that has the same functionality.

'write: false' will not write the article to the output path.

While it's not as descriptive as 'draft' or 'published' it does have
documentation that is discoverable and allows us to solve this issue
without adding more code complexity to the project.