Update 1-desktop.html.md
Fixed the incorrect usage of markdown inside of HTML tags, which wasn't displaying correctly on the web page. Also fixed the numbering system that was causing problems on the site. Note that the docs will probably look strange with this method of styling, as they already do, but I tested it on the site and it looked like expected.